requirements.txt
aiosqlite==0.22.1
argon2-cffi==25.1.0
bcrypt==5.0.0
email-validator==2.3.0
fastapi==0.128.1
Jinja2==3.1.6
passlib==1.7.4
pydantic==2.12.5
pydantic-settings==2.12.0
PyJWT==2.11.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
SQLAlchemy==2.0.46
uvicorn==0.40.0